move dashboardMainActivity to MVP UI package and remove useless code
